@@ -3,30 +3,35 @@
   <Import Project="..\..\common.props" />
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
     <BlazorWebAssemblyLoadAllGlobalizationData>true</BlazorWebAssemblyLoadAllGlobalizationData>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Blazorise.Bootstrap" Version="0.9.4.1" />
-    <PackageReference Include="Blazorise.Icons.FontAwesome" Version="0.9.4.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="5.0.*" />
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="5.0.*" />
+    <PackageReference Include="Blazorise.Bootstrap5" Version="1.3.2" />
+    <PackageReference Include="Blazorise.Icons.FontAwesome" Version="1.3.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="8.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="8.0.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Volo.Abp.Autofac.WebAssembly" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.AspNetCore.Components.WebAssembly.BasicTheme" Version="4.4.3" />
+   <PackageReference Include="Volo.Abp.AspNetCore.Components.WebAssembly.LeptonXLiteTheme" Version="3.0.*-*" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Volo.Abp.Identity.Blazor.WebAssembly" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.TenantManagement.Blazor.WebAssembly" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.SettingManagement.Blazor.WebAssembly" Version="4.4.3" />
+    <PackageReference Include="Volo.Abp.Autofac.WebAssembly" Version="8.0.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Volo.Abp.Identity.Blazor.WebAssembly" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.TenantManagement.Blazor.WebAssembly" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.SettingManagement.Blazor.WebAssembly" Version="8.0.2" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\AbpHideTenantSwitch.HttpApi.Client\AbpHideTenantSwitch.HttpApi.Client.csproj" />
   </ItemGroup>
+
 
 </Project>
