@@ -3,7 +3,8 @@
   <Import Project="..\..\common.props" />
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
     <RootNamespace>AbpHideTenantSwitch</RootNamespace>
   </PropertyGroup>
 
@@ -12,16 +13,16 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Volo.Abp.Emailing" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.Identity.Domain" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.PermissionManagement.Domain.Identity" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.IdentityServer.Domain" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.PermissionManagement.Domain.IdentityServer" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.BackgroundJobs.Domain" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.AuditLogging.Domain" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.TenantManagement.Domain" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.FeatureManagement.Domain" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.SettingManagement.Domain" Version="4.4.3" />
+    <PackageReference Include="Volo.Abp.Emailing" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.Identity.Domain" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.PermissionManagement.Domain.Identity" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.BackgroundJobs.Domain" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.AuditLogging.Domain" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.TenantManagement.Domain" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.FeatureManagement.Domain" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.SettingManagement.Domain" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.OpenIddict.Domain" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.PermissionManagement.Domain.OpenIddict" Version="8.0.2" />
   </ItemGroup>
 
 </Project>
