@@ -3,27 +3,28 @@
   <Import Project="..\..\common.props" />
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
     <RootNamespace>AbpHideTenantSwitch</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\AbpHideTenantSwitch.Domain\AbpHideTenantSwitch.Domain.csproj" />
-    <PackageReference Include="Volo.Abp.EntityFrameworkCore.SqlServer" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.PermissionManagement.EntityFrameworkCore" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.SettingManagement.EntityFrameworkCore" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.Identity.EntityFrameworkCore" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.IdentityServer.EntityFrameworkCore" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.BackgroundJobs.EntityFrameworkCore" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.AuditLogging.EntityFrameworkCore" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.TenantManagement.EntityFrameworkCore" Version="4.4.3" />
-    <PackageReference Include="Volo.Abp.FeatureManagement.EntityFrameworkCore" Version="4.4.3" />
+    <PackageReference Include="Volo.Abp.EntityFrameworkCore.SqlServer" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.PermissionManagement.EntityFrameworkCore" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.SettingManagement.EntityFrameworkCore" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.Identity.EntityFrameworkCore" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.BackgroundJobs.EntityFrameworkCore" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.AuditLogging.EntityFrameworkCore" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.TenantManagement.EntityFrameworkCore" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.FeatureManagement.EntityFrameworkCore" Version="8.0.2" />
+    <PackageReference Include="Volo.Abp.OpenIddict.EntityFrameworkCore" Version="8.0.2" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="5.0.*">
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="8.0.0">
+      <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-      <PrivateAssets>compile; contentFiles; build; buildMultitargeting; buildTransitive; analyzers; native</PrivateAssets>
     </PackageReference>
   </ItemGroup>
 
