@@ -2,16 +2,16 @@
 using Volo.Abp.ObjectExtending;
 using Volo.Abp.Threading;
 
-namespace AbpHideTenantSwitch
-{
-    public static class AbpHideTenantSwitchDtoExtensions
-    {
-        private static readonly OneTimeRunner OneTimeRunner = new OneTimeRunner();
+namespace AbpHideTenantSwitch;
 
-        public static void Configure()
+public static class AbpHideTenantSwitchDtoExtensions
+{
+    private static readonly OneTimeRunner OneTimeRunner = new OneTimeRunner();
+
+    public static void Configure()
+    {
+        OneTimeRunner.Run(() =>
         {
-            OneTimeRunner.Run(() =>
-            {
                 /* You can add extension properties to DTOs
                  * defined in the depended modules.
                  *
@@ -23,7 +23,6 @@ namespace AbpHideTenantSwitch
                  * See the documentation for more:
                  * https://docs.abp.io/en/abp/latest/Object-Extensions
                  */
-            });
-        }
+        });
     }
 }
