@@ -1,10 +1,9 @@
-﻿namespace AbpHideTenantSwitch.Permissions
-{
-    public static class AbpHideTenantSwitchPermissions
-    {
-        public const string GroupName = "AbpHideTenantSwitch";
+﻿namespace AbpHideTenantSwitch.Permissions;
 
-        //Add your own permission names. Example:
-        //public const string MyPermission1 = GroupName + ".MyPermission1";
-    }
+public static class AbpHideTenantSwitchPermissions
+{
+    public const string GroupName = "AbpHideTenantSwitch";
+
+    //Add your own permission names. Example:
+    //public const string MyPermission1 = GroupName + ".MyPermission1";
 }
