@@ -1,20 +1,19 @@
 ﻿using Volo.Abp.Bundling;
 
-namespace AbpHideTenantSwitch.Blazor
+namespace AbpHideTenantSwitch.Blazor;
+
+/* Add your global styles/scripts here.
+ * See https://docs.abp.io/en/abp/latest/UI/Blazor/Global-Scripts-Styles to learn how to use it
+ */
+public class AbpHideTenantSwitchBundleContributor : IBundleContributor
 {
-    /* Add your global styles/scripts here.
-     * See https://docs.abp.io/en/abp/latest/UI/Blazor/Global-Scripts-Styles to learn how to use it
-     */
-    public class AbpHideTenantSwitchBundleContributor : IBundleContributor
+    public void AddScripts(BundleContext context)
     {
-        public void AddScripts(BundleContext context)
-        {
 
-        }
+    }
 
-        public void AddStyles(BundleContext context)
-        {
-            context.Add("main.css", true);
-        }
+    public void AddStyles(BundleContext context)
+    {
+        context.Add("main.css", true);
     }
 }
