@@ -7,39 +7,35 @@ using Microsoft.Extensions.Logging;
 using Serilog;
 using Serilog.Events;
 
-namespace AbpHideTenantSwitch.DbMigrator
+namespace AbpHideTenantSwitch.DbMigrator;
+
+class Program
 {
-    class Program
+    static async Task Main(string[] args)
     {
-        static async Task Main(string[] args)
-        {
-            Log.Logger = new LoggerConfiguration()
-                .MinimumLevel.Information()
-                .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
-                .MinimumLevel.Override("Volo.Abp", LogEventLevel.Warning)
+        Log.Logger = new LoggerConfiguration()
+            .MinimumLevel.Information()
+            .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
+            .MinimumLevel.Override("Volo.Abp", LogEventLevel.Warning)
 #if DEBUG
                 .MinimumLevel.Override("AbpHideTenantSwitch", LogEventLevel.Debug)
 #else
                 .MinimumLevel.Override("AbpHideTenantSwitch", LogEventLevel.Information)
 #endif
                 .Enrich.FromLogContext()
-                .WriteTo.Async(c => c.File("Logs/logs.txt"))
-                .WriteTo.Async(c => c.Console())
-                .CreateLogger();
+            .WriteTo.Async(c => c.File("Logs/logs.txt"))
+            .WriteTo.Async(c => c.Console())
+            .CreateLogger();
 
-            await CreateHostBuilder(args).RunConsoleAsync();
-        }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureAppConfiguration(build =>
-                {
-                    build.AddJsonFile("appsettings.secrets.json", optional: true);
-                })
-                .ConfigureLogging((context, logging) => logging.ClearProviders())
-                .ConfigureServices((hostContext, services) =>
-                {
-                    services.AddHostedService<DbMigratorHostedService>();
-                });
+        await CreateHostBuilder(args).RunConsoleAsync();
     }
+
+    public static IHostBuilder CreateHostBuilder(string[] args) =>
+        Host.CreateDefaultBuilder(args)
+            .AddAppSettingsSecretsJson()
+            .ConfigureLogging((context, logging) => logging.ClearProviders())
+            .ConfigureServices((hostContext, services) =>
+            {
+                services.AddHostedService<DbMigratorHostedService>();
+            });
 }
