@@ -1,7 +1,6 @@
-﻿namespace AbpHideTenantSwitch
+﻿namespace AbpHideTenantSwitch;
+
+public static class AbpHideTenantSwitchDomainErrorCodes
 {
-    public static class AbpHideTenantSwitchDomainErrorCodes
-    {
-        /* You can add your business exception error codes here, as constants */
-    }
+    /* You can add your business exception error codes here, as constants */
 }
