@@ -1,15 +1,15 @@
 ﻿using Volo.Abp.Threading;
 
-namespace AbpHideTenantSwitch
-{
-    public static class AbpHideTenantSwitchGlobalFeatureConfigurator
-    {
-        private static readonly OneTimeRunner OneTimeRunner = new OneTimeRunner();
+namespace AbpHideTenantSwitch;
 
-        public static void Configure()
+public static class AbpHideTenantSwitchGlobalFeatureConfigurator
+{
+    private static readonly OneTimeRunner OneTimeRunner = new OneTimeRunner();
+
+    public static void Configure()
+    {
+        OneTimeRunner.Run(() =>
         {
-            OneTimeRunner.Run(() =>
-            {
                 /* You can configure (enable/disable) global features of the used modules here.
                  *
                  * YOU CAN SAFELY DELETE THIS CLASS AND REMOVE ITS USAGES IF YOU DON'T NEED TO IT!
@@ -17,7 +17,6 @@ namespace AbpHideTenantSwitch
                  * Please refer to the documentation to lear more about the Global Features System:
                  * https://docs.abp.io/en/abp/latest/Global-Features
                  */
-            });
-        }
+        });
     }
 }
