@@ -1,10 +1,9 @@
 ﻿using Volo.Abp.Localization;
 
-namespace AbpHideTenantSwitch.Localization
-{
-    [LocalizationResourceName("AbpHideTenantSwitch")]
-    public class AbpHideTenantSwitchResource
-    {
+namespace AbpHideTenantSwitch.Localization;
 
-    }
+[LocalizationResourceName("AbpHideTenantSwitch")]
+public class AbpHideTenantSwitchResource
+{
+
 }
