@@ -1,11 +1,10 @@
-﻿namespace AbpHideTenantSwitch.MultiTenancy
+﻿namespace AbpHideTenantSwitch.MultiTenancy;
+
+public static class MultiTenancyConsts
 {
-    public static class MultiTenancyConsts
-    {
-        /* Enable/disable multi-tenancy easily in a single point.
-         * If you will never need to multi-tenancy, you can remove
-         * related modules and code parts, including this file.
-         */
-        public const bool IsEnabled = true;
-    }
+    /* Enable/disable multi-tenancy easily in a single point.
+     * If you will never need to multi-tenancy, you can remove
+     * related modules and code parts, including this file.
+     */
+    public const bool IsEnabled = true;
 }
