@@ -1,9 +1,8 @@
-﻿namespace AbpHideTenantSwitch
-{
-    public static class AbpHideTenantSwitchConsts
-    {
-        public const string DbTablePrefix = "App";
+﻿namespace AbpHideTenantSwitch;
 
-        public const string DbSchema = null;
-    }
+public static class AbpHideTenantSwitchConsts
+{
+    public const string DbTablePrefix = "App";
+
+    public const string DbSchema = null;
 }
