@@ -13,208 +13,206 @@ using Volo.Abp.Identity;
 using Volo.Abp.MultiTenancy;
 using Volo.Abp.TenantManagement;
 
-namespace AbpHideTenantSwitch.Data
+namespace AbpHideTenantSwitch.Data;
+
+public class AbpHideTenantSwitchDbMigrationService : ITransientDependency
 {
-    public class AbpHideTenantSwitchDbMigrationService : ITransientDependency
+    public ILogger<AbpHideTenantSwitchDbMigrationService> Logger { get; set; }
+
+    private readonly IDataSeeder _dataSeeder;
+    private readonly IEnumerable<IAbpHideTenantSwitchDbSchemaMigrator> _dbSchemaMigrators;
+    private readonly ITenantRepository _tenantRepository;
+    private readonly ICurrentTenant _currentTenant;
+
+    public AbpHideTenantSwitchDbMigrationService(
+        IDataSeeder dataSeeder,
+        IEnumerable<IAbpHideTenantSwitchDbSchemaMigrator> dbSchemaMigrators,
+        ITenantRepository tenantRepository,
+        ICurrentTenant currentTenant)
     {
-        public ILogger<AbpHideTenantSwitchDbMigrationService> Logger { get; set; }
+        _dataSeeder = dataSeeder;
+        _dbSchemaMigrators = dbSchemaMigrators;
+        _tenantRepository = tenantRepository;
+        _currentTenant = currentTenant;
 
-        private readonly IDataSeeder _dataSeeder;
-        private readonly IEnumerable<IAbpHideTenantSwitchDbSchemaMigrator> _dbSchemaMigrators;
-        private readonly ITenantRepository _tenantRepository;
-        private readonly ICurrentTenant _currentTenant;
+        Logger = NullLogger<AbpHideTenantSwitchDbMigrationService>.Instance;
+    }
 
-        public AbpHideTenantSwitchDbMigrationService(
-            IDataSeeder dataSeeder,
-            IEnumerable<IAbpHideTenantSwitchDbSchemaMigrator> dbSchemaMigrators,
-            ITenantRepository tenantRepository,
-            ICurrentTenant currentTenant)
+    public async Task MigrateAsync()
+    {
+        var initialMigrationAdded = AddInitialMigrationIfNotExist();
+
+        if (initialMigrationAdded)
         {
-            _dataSeeder = dataSeeder;
-            _dbSchemaMigrators = dbSchemaMigrators;
-            _tenantRepository = tenantRepository;
-            _currentTenant = currentTenant;
-
-            Logger = NullLogger<AbpHideTenantSwitchDbMigrationService>.Instance;
+            return;
         }
 
-        public async Task MigrateAsync()
+        Logger.LogInformation("Started database migrations...");
+
+        await MigrateDatabaseSchemaAsync();
+        await SeedDataAsync();
+
+        Logger.LogInformation($"Successfully completed host database migrations.");
+
+        var tenants = await _tenantRepository.GetListAsync(includeDetails: true);
+
+        var migratedDatabaseSchemas = new HashSet<string>();
+        foreach (var tenant in tenants)
         {
-            var initialMigrationAdded = AddInitialMigrationIfNotExist();
-
-            if (initialMigrationAdded)
+            using (_currentTenant.Change(tenant.Id))
             {
-                return;
-            }
-
-            Logger.LogInformation("Started database migrations...");
-
-            await MigrateDatabaseSchemaAsync();
-            await SeedDataAsync();
-
-            Logger.LogInformation($"Successfully completed host database migrations.");
-
-            var tenants = await _tenantRepository.GetListAsync(includeDetails: true);
-
-            var migratedDatabaseSchemas = new HashSet<string>();
-            foreach (var tenant in tenants)
-            {
-                using (_currentTenant.Change(tenant.Id))
+                if (tenant.ConnectionStrings.Any())
                 {
-                    if (tenant.ConnectionStrings.Any())
+                    var tenantConnectionStrings = tenant.ConnectionStrings
+                        .Select(x => x.Value)
+                        .ToList();
+
+                    if (!migratedDatabaseSchemas.IsSupersetOf(tenantConnectionStrings))
                     {
-                        var tenantConnectionStrings = tenant.ConnectionStrings
-                            .Select(x => x.Value)
-                            .ToList();
+                        await MigrateDatabaseSchemaAsync(tenant);
 
-                        if (!migratedDatabaseSchemas.IsSupersetOf(tenantConnectionStrings))
-                        {
-                            await MigrateDatabaseSchemaAsync(tenant);
-
-                            migratedDatabaseSchemas.AddIfNotContains(tenantConnectionStrings);
-                        }
+                        migratedDatabaseSchemas.AddIfNotContains(tenantConnectionStrings);
                     }
-
-                    await SeedDataAsync(tenant);
                 }
 
-                Logger.LogInformation($"Successfully completed {tenant.Name} tenant database migrations.");
+                await SeedDataAsync(tenant);
             }
 
-            Logger.LogInformation("Successfully completed all database migrations.");
-            Logger.LogInformation("You can safely end this process...");
+            Logger.LogInformation($"Successfully completed {tenant.Name} tenant database migrations.");
         }
 
-        private async Task MigrateDatabaseSchemaAsync(Tenant tenant = null)
-        {
-            Logger.LogInformation(
-                $"Migrating schema for {(tenant == null ? "host" : tenant.Name + " tenant")} database...");
+        Logger.LogInformation("Successfully completed all database migrations.");
+        Logger.LogInformation("You can safely end this process...");
+    }
 
-            foreach (var migrator in _dbSchemaMigrators)
-            {
-                await migrator.MigrateAsync();
-            }
+    private async Task MigrateDatabaseSchemaAsync(Tenant? tenant = null)
+    {
+        Logger.LogInformation(
+            $"Migrating schema for {(tenant == null ? "host" : tenant.Name + " tenant")} database...");
+
+        foreach (var migrator in _dbSchemaMigrators)
+        {
+            await migrator.MigrateAsync();
         }
+    }
 
-        private async Task SeedDataAsync(Tenant tenant = null)
+    private async Task SeedDataAsync(Tenant? tenant = null)
+    {
+        Logger.LogInformation($"Executing {(tenant == null ? "host" : tenant.Name + " tenant")} database seed...");
+
+        await _dataSeeder.SeedAsync(new DataSeedContext(tenant?.Id)
+            .WithProperty(IdentityDataSeedContributor.AdminEmailPropertyName, IdentityDataSeedContributor.AdminEmailDefaultValue)
+            .WithProperty(IdentityDataSeedContributor.AdminPasswordPropertyName, IdentityDataSeedContributor.AdminPasswordDefaultValue)
+        );
+    }
+
+    private bool AddInitialMigrationIfNotExist()
+    {
+        try
         {
-            Logger.LogInformation($"Executing {(tenant == null ? "host" : tenant.Name + " tenant")} database seed...");
-
-            await _dataSeeder.SeedAsync(new DataSeedContext(tenant?.Id)
-                .WithProperty(IdentityDataSeedContributor.AdminEmailPropertyName, IdentityDataSeedContributor.AdminEmailDefaultValue)
-                .WithProperty(IdentityDataSeedContributor.AdminPasswordPropertyName, IdentityDataSeedContributor.AdminPasswordDefaultValue)
-            );
-        }
-
-        private bool AddInitialMigrationIfNotExist()
-        {
-            try
-            {
-                if (!DbMigrationsProjectExists())
-                {
-                    return false;
-                }
-            }
-            catch (Exception)
+            if (!DbMigrationsProjectExists())
             {
                 return false;
             }
-
-            try
-            {
-                if (!MigrationsFolderExists())
-                {
-                    AddInitialMigration();
-                    return true;
-                }
-                else
-                {
-                    return false;
-                }
-            }
-            catch (Exception e)
-            {
-                Logger.LogWarning("Couldn't determinate if any migrations exist : " + e.Message);
-                return false;
-            }
+        }
+        catch (Exception)
+        {
+            return false;
         }
 
-        private bool DbMigrationsProjectExists()
+        try
         {
-            var dbMigrationsProjectFolder = GetEntityFrameworkCoreProjectFolderPath();
-
-            return dbMigrationsProjectFolder != null;
-        }
-
-        private bool MigrationsFolderExists()
-        {
-            var dbMigrationsProjectFolder = GetEntityFrameworkCoreProjectFolderPath();
-
-            return Directory.Exists(Path.Combine(dbMigrationsProjectFolder, "Migrations"));
-        }
-
-        private void AddInitialMigration()
-        {
-            Logger.LogInformation("Creating initial migration...");
-
-            string argumentPrefix;
-            string fileName;
-
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX) || RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+            if (!MigrationsFolderExists())
             {
-                argumentPrefix = "-c";
-                fileName = "/bin/bash";
+                AddInitialMigration();
+                return true;
             }
             else
             {
-                argumentPrefix = "/C";
-                fileName = "cmd.exe";
-            }
-
-            var procStartInfo = new ProcessStartInfo(fileName,
-                $"{argumentPrefix} \"abp create-migration-and-run-migrator \"{GetEntityFrameworkCoreProjectFolderPath()}\"\""
-            );
-
-            try
-            {
-                Process.Start(procStartInfo);
-            }
-            catch (Exception)
-            {
-                throw new Exception("Couldn't run ABP CLI...");
+                return false;
             }
         }
-
-        private string GetEntityFrameworkCoreProjectFolderPath()
+        catch (Exception e)
         {
-            var slnDirectoryPath = GetSolutionDirectoryPath();
-
-            if (slnDirectoryPath == null)
-            {
-                throw new Exception("Solution folder not found!");
-            }
-
-            var srcDirectoryPath = Path.Combine(slnDirectoryPath, "src");
-
-            return Directory.GetDirectories(srcDirectoryPath)
-                .FirstOrDefault(d => d.EndsWith(".EntityFrameworkCore"));
+            Logger.LogWarning("Couldn't determinate if any migrations exist : " + e.Message);
+            return false;
         }
+    }
 
-        private string GetSolutionDirectoryPath()
+    private bool DbMigrationsProjectExists()
+    {
+        var dbMigrationsProjectFolder = GetEntityFrameworkCoreProjectFolderPath();
+
+        return dbMigrationsProjectFolder != null;
+    }
+
+    private bool MigrationsFolderExists()
+    {
+        var dbMigrationsProjectFolder = GetEntityFrameworkCoreProjectFolderPath();
+        return dbMigrationsProjectFolder != null && Directory.Exists(Path.Combine(dbMigrationsProjectFolder, "Migrations"));
+    }
+
+    private void AddInitialMigration()
+    {
+        Logger.LogInformation("Creating initial migration...");
+
+        string argumentPrefix;
+        string fileName;
+
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX) || RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
         {
-            var currentDirectory = new DirectoryInfo(Directory.GetCurrentDirectory());
-
-            while (Directory.GetParent(currentDirectory.FullName) != null)
-            {
-                currentDirectory = Directory.GetParent(currentDirectory.FullName);
-
-                if (Directory.GetFiles(currentDirectory.FullName).FirstOrDefault(f => f.EndsWith(".sln")) != null)
-                {
-                    return currentDirectory.FullName;
-                }
-            }
-
-            return null;
+            argumentPrefix = "-c";
+            fileName = "/bin/bash";
         }
+        else
+        {
+            argumentPrefix = "/C";
+            fileName = "cmd.exe";
+        }
+
+        var procStartInfo = new ProcessStartInfo(fileName,
+            $"{argumentPrefix} \"abp create-migration-and-run-migrator \"{GetEntityFrameworkCoreProjectFolderPath()}\"\""
+        );
+
+        try
+        {
+            Process.Start(procStartInfo);
+        }
+        catch (Exception)
+        {
+            throw new Exception("Couldn't run ABP CLI...");
+        }
+    }
+
+    private string? GetEntityFrameworkCoreProjectFolderPath()
+    {
+        var slnDirectoryPath = GetSolutionDirectoryPath();
+
+        if (slnDirectoryPath == null)
+        {
+            throw new Exception("Solution folder not found!");
+        }
+
+        var srcDirectoryPath = Path.Combine(slnDirectoryPath, "src");
+
+        return Directory.GetDirectories(srcDirectoryPath)
+            .FirstOrDefault(d => d.EndsWith(".EntityFrameworkCore"));
+    }
+
+    private string? GetSolutionDirectoryPath()
+    {
+        var currentDirectory = new DirectoryInfo(Directory.GetCurrentDirectory());
+
+        while (currentDirectory != null && Directory.GetParent(currentDirectory.FullName) != null)
+        {
+            currentDirectory = Directory.GetParent(currentDirectory.FullName);
+
+            if (currentDirectory != null && Directory.GetFiles(currentDirectory.FullName).FirstOrDefault(f => f.EndsWith(".sln")) != null)
+            {
+                return currentDirectory.FullName;
+            }
+        }
+
+        return null;
     }
 }
