@@ -1,10 +1,9 @@
-﻿namespace AbpHideTenantSwitch.Settings
-{
-    public static class AbpHideTenantSwitchSettings
-    {
-        private const string Prefix = "AbpHideTenantSwitch";
+﻿namespace AbpHideTenantSwitch.Settings;
 
-        //Add your own setting names here. Example:
-        //public const string MySetting1 = Prefix + ".MySetting1";
-    }
+public static class AbpHideTenantSwitchSettings
+{
+    private const string Prefix = "AbpHideTenantSwitch";
+
+    //Add your own setting names here. Example:
+    //public const string MySetting1 = Prefix + ".MySetting1";
 }
