@@ -7,80 +7,80 @@ using Volo.Abp.EntityFrameworkCore;
 using Volo.Abp.FeatureManagement.EntityFrameworkCore;
 using Volo.Abp.Identity;
 using Volo.Abp.Identity.EntityFrameworkCore;
-using Volo.Abp.IdentityServer.EntityFrameworkCore;
+using Volo.Abp.OpenIddict.EntityFrameworkCore;
 using Volo.Abp.PermissionManagement.EntityFrameworkCore;
 using Volo.Abp.SettingManagement.EntityFrameworkCore;
 using Volo.Abp.TenantManagement;
 using Volo.Abp.TenantManagement.EntityFrameworkCore;
 
-namespace AbpHideTenantSwitch.EntityFrameworkCore
+namespace AbpHideTenantSwitch.EntityFrameworkCore;
+
+[ReplaceDbContext(typeof(IIdentityDbContext))]
+[ReplaceDbContext(typeof(ITenantManagementDbContext))]
+[ConnectionStringName("Default")]
+public class AbpHideTenantSwitchDbContext :
+    AbpDbContext<AbpHideTenantSwitchDbContext>,
+    IIdentityDbContext,
+    ITenantManagementDbContext
 {
-    [ReplaceDbContext(typeof(IIdentityDbContext))]
-    [ReplaceDbContext(typeof(ITenantManagementDbContext))]
-    [ConnectionStringName("Default")]
-    public class AbpHideTenantSwitchDbContext : 
-        AbpDbContext<AbpHideTenantSwitchDbContext>,
-        IIdentityDbContext,
-        ITenantManagementDbContext
+    /* Add DbSet properties for your Aggregate Roots / Entities here. */
+
+    #region Entities from the modules
+
+    /* Notice: We only implemented IIdentityDbContext and ITenantManagementDbContext
+     * and replaced them for this DbContext. This allows you to perform JOIN
+     * queries for the entities of these modules over the repositories easily. You
+     * typically don't need that for other modules. But, if you need, you can
+     * implement the DbContext interface of the needed module and use ReplaceDbContext
+     * attribute just like IIdentityDbContext and ITenantManagementDbContext.
+     *
+     * More info: Replacing a DbContext of a module ensures that the related module
+     * uses this DbContext on runtime. Otherwise, it will use its own DbContext class.
+     */
+
+    //Identity
+    public DbSet<IdentityUser> Users { get; set; }
+    public DbSet<IdentityRole> Roles { get; set; }
+    public DbSet<IdentityClaimType> ClaimTypes { get; set; }
+    public DbSet<OrganizationUnit> OrganizationUnits { get; set; }
+    public DbSet<IdentitySecurityLog> SecurityLogs { get; set; }
+    public DbSet<IdentityLinkUser> LinkUsers { get; set; }
+    public DbSet<IdentityUserDelegation> UserDelegations { get; set; }
+
+    // Tenant Management
+    public DbSet<Tenant> Tenants { get; set; }
+    public DbSet<TenantConnectionString> TenantConnectionStrings { get; set; }
+
+    #endregion
+
+    public AbpHideTenantSwitchDbContext(DbContextOptions<AbpHideTenantSwitchDbContext> options)
+        : base(options)
     {
-        /* Add DbSet properties for your Aggregate Roots / Entities here. */
-        
-        #region Entities from the modules
-        
-        /* Notice: We only implemented IIdentityDbContext and ITenantManagementDbContext
-         * and replaced them for this DbContext. This allows you to perform JOIN
-         * queries for the entities of these modules over the repositories easily. You
-         * typically don't need that for other modules. But, if you need, you can
-         * implement the DbContext interface of the needed module and use ReplaceDbContext
-         * attribute just like IIdentityDbContext and ITenantManagementDbContext.
-         *
-         * More info: Replacing a DbContext of a module ensures that the related module
-         * uses this DbContext on runtime. Otherwise, it will use its own DbContext class.
-         */
-        
-        //Identity
-        public DbSet<IdentityUser> Users { get; set; }
-        public DbSet<IdentityRole> Roles { get; set; }
-        public DbSet<IdentityClaimType> ClaimTypes { get; set; }
-        public DbSet<OrganizationUnit> OrganizationUnits { get; set; }
-        public DbSet<IdentitySecurityLog> SecurityLogs { get; set; }
-        public DbSet<IdentityLinkUser> LinkUsers { get; set; }
-        
-        // Tenant Management
-        public DbSet<Tenant> Tenants { get; set; }
-        public DbSet<TenantConnectionString> TenantConnectionStrings { get; set; }
 
-        #endregion
-        
-        public AbpHideTenantSwitchDbContext(DbContextOptions<AbpHideTenantSwitchDbContext> options)
-            : base(options)
-        {
+    }
 
-        }
+    protected override void OnModelCreating(ModelBuilder builder)
+    {
+        base.OnModelCreating(builder);
 
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            base.OnModelCreating(builder);
+        /* Include modules to your migration db context */
 
-            /* Include modules to your migration db context */
+        builder.ConfigurePermissionManagement();
+        builder.ConfigureSettingManagement();
+        builder.ConfigureBackgroundJobs();
+        builder.ConfigureAuditLogging();
+        builder.ConfigureIdentity();
+        builder.ConfigureOpenIddict();
+        builder.ConfigureFeatureManagement();
+        builder.ConfigureTenantManagement();
 
-            builder.ConfigurePermissionManagement();
-            builder.ConfigureSettingManagement();
-            builder.ConfigureBackgroundJobs();
-            builder.ConfigureAuditLogging();
-            builder.ConfigureIdentity();
-            builder.ConfigureIdentityServer();
-            builder.ConfigureFeatureManagement();
-            builder.ConfigureTenantManagement();
+        /* Configure your own tables/entities inside here */
 
-            /* Configure your own tables/entities inside here */
-
-            //builder.Entity<YourEntity>(b =>
-            //{
-            //    b.ToTable(AbpHideTenantSwitchConsts.DbTablePrefix + "YourEntities", AbpHideTenantSwitchConsts.DbSchema);
-            //    b.ConfigureByConvention(); //auto configure for the base class props
-            //    //...
-            //});
-        }
+        //builder.Entity<YourEntity>(b =>
+        //{
+        //    b.ToTable(AbpHideTenantSwitchConsts.DbTablePrefix + "YourEntities", AbpHideTenantSwitchConsts.DbSchema);
+        //    b.ConfigureByConvention(); //auto configure for the base class props
+        //    //...
+        //});
     }
 }
