@@ -3,19 +3,19 @@ using Volo.Abp.Identity;
 using Volo.Abp.ObjectExtending;
 using Volo.Abp.Threading;
 
-namespace AbpHideTenantSwitch.EntityFrameworkCore
+namespace AbpHideTenantSwitch.EntityFrameworkCore;
+
+public static class AbpHideTenantSwitchEfCoreEntityExtensionMappings
 {
-    public static class AbpHideTenantSwitchEfCoreEntityExtensionMappings
+    private static readonly OneTimeRunner OneTimeRunner = new OneTimeRunner();
+
+    public static void Configure()
     {
-        private static readonly OneTimeRunner OneTimeRunner = new OneTimeRunner();
+        AbpHideTenantSwitchGlobalFeatureConfigurator.Configure();
+        AbpHideTenantSwitchModuleExtensionConfigurator.Configure();
 
-        public static void Configure()
+        OneTimeRunner.Run(() =>
         {
-            AbpHideTenantSwitchGlobalFeatureConfigurator.Configure();
-            AbpHideTenantSwitchModuleExtensionConfigurator.Configure();
-
-            OneTimeRunner.Run(() =>
-            {
                 /* You can configure extra properties for the
                  * entities defined in the modules used by your application.
                  *
@@ -39,7 +39,6 @@ namespace AbpHideTenantSwitch.EntityFrameworkCore
                  * See the documentation for more:
                  * https://docs.abp.io/en/abp/latest/Customizing-Application-Modules-Extending-Entities
                  */
-            });
-        }
+        });
     }
 }
