@@ -1,7 +1,6 @@
 ﻿@page "/"
 @using Volo.Abp.MultiTenancy
 @inherits AbpHideTenantSwitchComponentBase
-@inject ICurrentTenant CurrentTenant
 @inject AuthenticationStateProvider AuthenticationStateProvider
 <div class="container">
     <div class="p-5 text-center">
@@ -21,6 +20,31 @@
         }
     </div>
 
+    <div class="card">
+        <div class="card-body">
+            <div class="row">
+                <div class="col-md-auto text-center">
+                    <img src="https://abp.io/assets/png/mastering-abp-framework.webp" style="max-width: 400px;" class="w-100 mb-5 my-md-3">
+                </div>
+                <div class="col-md d-flex align-items-center">
+                    <div class="pe-0 pe-md-4">
+                        <small class="text-uppercase text-muted">THE OFFICIAL GUIDE</small>
+                        <h2 class="mb-4">Mastering ABP Framework</h2>
+                        <p class="mb-4">Written by the creator of the ABP Framework, this book will help you gain a complete understanding of the framework and modern web application development techniques.</p>
+                        <div class="mb-4">
+                            <a href="https://www.amazon.com/gp/product/B097Z2DM8Q/ref=dbs_a_def_rwt_hsch_vapi_tkin_p1_i0" class="btn btn-success mb-1">
+                                Buy on Amazon US
+                            </a>
+                            <a href="https://www.packtpub.com/product/mastering-abp-framework/9781801079242" class="btn btn-primary mb-1">
+                                Buy on PACKT
+                            </a>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+
     <div class="my-3 text-center">
         <h3>Let's improve your application!</h3>
         <p>Here are some links to help you get started:</p>
@@ -36,14 +60,14 @@
                         <a href="https://docs.abp.io/en/abp/latest?ref=tmpl" target="_blank" class="btn btn-link px-1">See Documents <i class="fas fa-chevron-right"></i></a>
                     </div>
                 </div>
-                <div class="col-lg-4 border-left">
+                <div class="col-lg-4 border-start">
                     <div class="p-4">
                         <h5 class="mb-3"><i class="fas fa-cubes text-secondary d-block my-3 fa-2x"></i> Samples</h5>
                         <p>See the example projects built with the ABP Framework.</p>
                         <a href="https://docs.abp.io/en/abp/latest/Samples/Index?ref=tmpl" target="_blank" class="btn btn-link px-1">All samples <i class="fas fa-chevron-right"></i></a>
                     </div>
                 </div>
-                <div class="col-lg-4 border-left">
+                <div class="col-lg-4 border-start">
                     <div class="p-4">
                         <h5 class="mb-3"><i class="fas fa-users text-secondary d-block my-3 fa-2x"></i> ABP Community</h5>
                         <p>Get involved with a vibrant community and become a contributor.</p>
@@ -60,7 +84,7 @@
                         <a href="https://blog.abp.io/abp?ref=tmpl" target="_blank" class="btn btn-link px-1">See Blog <i class="fas fa-chevron-right"></i></a>
                     </div>
                 </div>
-                <div class="col-lg-4 border-left">
+                <div class="col-lg-4 border-start">
                     <div class="p-4">
                         <h5 class="mb-3"><i class="fab fa-github text-secondary d-block my-3 fa-2x"></i> Github</h5>
                         <p>Do you love the ABP Framework? Please <strong>give a star</strong> to support it!</p>
@@ -69,7 +93,7 @@
 
                     </div>
                 </div>
-                <div class="col-lg-4 border-left">
+                <div class="col-lg-4 border-start">
                     <div class="p-4">
                         <h5 class="mb-3"><i class="fab fa-stack-overflow text-secondary d-block my-3 fa-2x"></i> Stackoverflow</h5>
                         <p>See answers to previously asked questions or ask a new one.</p>
@@ -102,7 +126,7 @@
                         </h6>
                     </div>
                 </div>
-                <div class="col-lg-2 border-left">
+                <div class="col-lg-2 border-start">
                     <div class="p-3">
                         <h6>
                             <i class="fas fa-plus d-block mb-3 fa- 2x text-secondary"></i> Application Modules
@@ -110,7 +134,7 @@
                         </h6>
                     </div>
                 </div>
-                <div class="col-lg-2 border-left">
+                <div class="col-lg-2 border-start">
                     <div class="p-3">
                         <h6>
                             <i class="fas fa-plus d-block mb-3 fa- 2x text-secondary"></i> Developer<br />Tools
@@ -118,7 +142,7 @@
                         </h6>
                     </div>
                 </div>
-                <div class="col-lg-2 border-left">
+                <div class="col-lg-2 border-start">
                     <div class="p-3">
                         <h6>
                             <i class="fas fa-plus d-block mb-3 fa- 2x text-secondary"></i> UI<br /> Themes
@@ -126,7 +150,7 @@
                         </h6>
                     </div>
                 </div>
-                <div class="col-lg-2 border-left">
+                <div class="col-lg-2 border-start">
                     <div class="p-3">
                         <h6>
                             <i class="fas fa-plus d-block mb-3 fa- 2x text-secondary"></i> Premium Support
@@ -134,7 +158,7 @@
                         </h6>
                     </div>
                 </div>
-                <div class="col-lg-2 border-left">
+                <div class="col-lg-2 border-start">
                     <div class="p-3">
                         <h6>
                             <i class="fas fa-plus d-block mb-3 fa- 2x text-secondary"></i> Additional Services
