@@ -1,52 +1,54 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using System;
+using Microsoft.Extensions.DependencyInjection;
+using Volo.Abp.Uow;
 using Volo.Abp.AuditLogging.EntityFrameworkCore;
 using Volo.Abp.BackgroundJobs.EntityFrameworkCore;
 using Volo.Abp.EntityFrameworkCore;
 using Volo.Abp.EntityFrameworkCore.SqlServer;
 using Volo.Abp.FeatureManagement.EntityFrameworkCore;
 using Volo.Abp.Identity.EntityFrameworkCore;
-using Volo.Abp.IdentityServer.EntityFrameworkCore;
 using Volo.Abp.Modularity;
+using Volo.Abp.OpenIddict.EntityFrameworkCore;
 using Volo.Abp.PermissionManagement.EntityFrameworkCore;
 using Volo.Abp.SettingManagement.EntityFrameworkCore;
 using Volo.Abp.TenantManagement.EntityFrameworkCore;
 
-namespace AbpHideTenantSwitch.EntityFrameworkCore
-{
-    [DependsOn(
-        typeof(AbpHideTenantSwitchDomainModule),
-        typeof(AbpIdentityEntityFrameworkCoreModule),
-        typeof(AbpIdentityServerEntityFrameworkCoreModule),
-        typeof(AbpPermissionManagementEntityFrameworkCoreModule),
-        typeof(AbpSettingManagementEntityFrameworkCoreModule),
-        typeof(AbpEntityFrameworkCoreSqlServerModule),
-        typeof(AbpBackgroundJobsEntityFrameworkCoreModule),
-        typeof(AbpAuditLoggingEntityFrameworkCoreModule),
-        typeof(AbpTenantManagementEntityFrameworkCoreModule),
-        typeof(AbpFeatureManagementEntityFrameworkCoreModule)
-        )]
-    public class AbpHideTenantSwitchEntityFrameworkCoreModule : AbpModule
-    {
-        public override void PreConfigureServices(ServiceConfigurationContext context)
-        {
-            AbpHideTenantSwitchEfCoreEntityExtensionMappings.Configure();
-        }
+namespace AbpHideTenantSwitch.EntityFrameworkCore;
 
-        public override void ConfigureServices(ServiceConfigurationContext context)
+[DependsOn(
+    typeof(AbpHideTenantSwitchDomainModule),
+    typeof(AbpIdentityEntityFrameworkCoreModule),
+    typeof(AbpOpenIddictEntityFrameworkCoreModule),
+    typeof(AbpPermissionManagementEntityFrameworkCoreModule),
+    typeof(AbpSettingManagementEntityFrameworkCoreModule),
+    typeof(AbpEntityFrameworkCoreSqlServerModule),
+    typeof(AbpBackgroundJobsEntityFrameworkCoreModule),
+    typeof(AbpAuditLoggingEntityFrameworkCoreModule),
+    typeof(AbpTenantManagementEntityFrameworkCoreModule),
+    typeof(AbpFeatureManagementEntityFrameworkCoreModule)
+    )]
+public class AbpHideTenantSwitchEntityFrameworkCoreModule : AbpModule
+{
+    public override void PreConfigureServices(ServiceConfigurationContext context)
+    {
+        AbpHideTenantSwitchEfCoreEntityExtensionMappings.Configure();
+    }
+
+    public override void ConfigureServices(ServiceConfigurationContext context)
+    {
+        context.Services.AddAbpDbContext<AbpHideTenantSwitchDbContext>(options =>
         {
-            context.Services.AddAbpDbContext<AbpHideTenantSwitchDbContext>(options =>
-            {
                 /* Remove "includeAllEntities: true" to create
                  * default repositories only for aggregate roots */
-                options.AddDefaultRepositories(includeAllEntities: true);
-            });
+            options.AddDefaultRepositories(includeAllEntities: true);
+        });
 
-            Configure<AbpDbContextOptions>(options =>
-            {
+        Configure<AbpDbContextOptions>(options =>
+        {
                 /* The main point to change your DBMS.
                  * See also AbpHideTenantSwitchMigrationsDbContextFactory for EF Core tooling. */
-                options.UseSqlServer();
-            });
-        }
+            options.UseSqlServer();
+        });
+
     }
 }
