@@ -5,16 +5,18 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 using Volo.Abp.Account.Web;
 using Volo.Abp.Account.Web.Pages.Account;
+using Volo.Abp.Identity;
 using Volo.Abp.TenantManagement;
 using IdentityUser = Volo.Abp.Identity.IdentityUser;
 
-namespace AbpHideTenantSwitch.HttpApi.Host.Pages.Account
+namespace AbpHideTenantSwitch.Pages.Account
 {
     public class CustomLoginModel : LoginModel
     {
         private readonly ITenantRepository _tenantRepository;
 
-        public CustomLoginModel(IAuthenticationSchemeProvider schemeProvider, IOptions<AbpAccountOptions> accountOptions, IOptions<IdentityOptions> identityOptions, ITenantRepository tenantRepository) : base(schemeProvider, accountOptions, identityOptions)
+        public CustomLoginModel(IAuthenticationSchemeProvider schemeProvider, IOptions<AbpAccountOptions> accountOptions, IOptions<IdentityOptions> identityOptions, ITenantRepository tenantRepository, IdentityDynamicClaimsPrincipalContributorCache contributorCache) 
+            : base(schemeProvider, accountOptions, identityOptions, contributorCache)
         {
             _tenantRepository = tenantRepository;
         }
@@ -55,7 +57,6 @@ namespace AbpHideTenantSwitch.HttpApi.Host.Pages.Account
                     }
                 }
             }
-
             return null;
         }
     }
