@@ -72,36 +72,83 @@ var abp = abp || {};
     /* LOCALIZATION ***********************************************/
 
     abp.localization = abp.localization || {};
+    abp.localization.internal = abp.localization.internal || {};
+    abp.localization.values =  abp.localization.values || {};
+    abp.localization.resources =  abp.localization.resources || {};
 
-    abp.localization.values = {};
-
-    abp.localization.localize = function (key, sourceName) {
-        if (sourceName === '_') { //A convention to suppress the localization
-            return key;
+    abp.localization.internal.getResource = function (resourceName) {
+        var resource = abp.localization.resources[resourceName];
+        if (resource) {
+            return resource;
+        }
+        
+        var legacySource = abp.localization.values[resourceName];
+        if (legacySource) {
+            return {
+                texts: abp.localization.values[resourceName],
+                baseResources: []
+            };
+        }
+        
+        abp.log.warn('Could not find localization source: ' + resourceName);        
+        return null;
+    };
+    
+    abp.localization.internal.localize = function (key, sourceName) {
+        var resource = abp.localization.internal.getResource(sourceName);
+        if (!resource){
+            return {
+                value: key,
+                found: false
+            };
         }
 
-        sourceName = sourceName || abp.localization.defaultResourceName;
-        if (!sourceName) {
-            abp.log.warn('Localization source name is not specified and the defaultResourceName was not defined!');
-            return key;
-        }
+        var value = resource.texts[key];
+        if (value === undefined) {            
+            for (var i = 0; i < resource.baseResources.length; i++){
+                var basedArguments = Array.prototype.slice.call(arguments, 0);
+                basedArguments[1] = resource.baseResources[i];
 
-        var source = abp.localization.values[sourceName];
-        if (!source) {
-            abp.log.warn('Could not find localization source: ' + sourceName);
-            return key;
-        }
-
-        var value = source[key];
-        if (value == undefined) {
-            return key;
+                var result = abp.localization.internal.localize.apply(this, basedArguments);
+                if (result.found){
+                    return result;
+                }
+            }
+            
+            return {
+                value: key,
+                found: false
+            };
         }
 
         var copiedArguments = Array.prototype.slice.call(arguments, 0);
         copiedArguments.splice(1, 1);
         copiedArguments[0] = value;
 
-        return abp.utils.formatString.apply(this, copiedArguments);
+        return {
+            value: abp.utils.formatString.apply(this, copiedArguments),
+            found: true
+        };
+    };
+
+    abp.localization.localize = function (key, sourceName) {
+        if (sourceName === '_') { //A convention to suppress the localization
+            return key;
+        }
+        
+        if (sourceName) {
+            return abp.localization.internal.localize.apply(this, arguments).value;
+        }
+
+        if (!abp.localization.defaultResourceName) {
+            abp.log.warn('Localization source name is not specified and the defaultResourceName was not defined!');
+            return key;
+        }
+
+        var copiedArguments = Array.prototype.slice.call(arguments, 0);
+        copiedArguments.splice(1, 1, abp.localization.defaultResourceName);
+
+        return abp.localization.internal.localize.apply(this, copiedArguments).value;
     };
 
     abp.localization.isLocalized = function (key, sourceName) {
@@ -114,17 +161,7 @@ var abp = abp || {};
             return false;
         }
 
-        var source = abp.localization.values[sourceName];
-        if (!source) {
-            return false;
-        }
-
-        var value = source[key];
-        if (value === undefined) {
-            return false;
-        }
-
-        return true;
+        return abp.localization.internal.localize(key, sourceName).found;
     };
 
     abp.localization.getResource = function (name) {
@@ -173,12 +210,10 @@ var abp = abp || {};
 
     abp.auth = abp.auth || {};
 
-    abp.auth.policies = abp.auth.policies || {};
-
     abp.auth.grantedPolicies = abp.auth.grantedPolicies || {};
 
     abp.auth.isGranted = function (policyName) {
-        return abp.auth.policies[policyName] != undefined && abp.auth.grantedPolicies[policyName] != undefined;
+        return abp.auth.grantedPolicies[policyName] != undefined;
     };
 
     abp.auth.isAnyGranted = function () {
@@ -372,7 +407,9 @@ var abp = abp || {};
             setTimeout(function () {
                 if (element) {
                     element.classList.remove('abp-block-area-disappearing');
-                    element.parentElement.removeChild(element);
+                    if (element.parentElement) {
+                        element.parentElement.removeChild(element);
+                    }
                 }
             }, 250);
         }
@@ -685,7 +722,7 @@ var abp = abp || {};
     }
 
     /**
-     * Escape HTML to help prevent XSS attacks. 
+     * Escape HTML to help prevent XSS attacks.
      */
     abp.utils.htmlEscape = function (html) {
         return typeof html === 'string' ? html.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;') : html;
@@ -757,7 +794,7 @@ var abp = abp || {};
             return toUtc(date);
         }
     };
-    
+
     /* FEATURES *************************************************/
 
     abp.features = abp.features || {};
@@ -772,5 +809,15 @@ var abp = abp || {};
     abp.features.get = function (name) {
         return abp.features.values[name];
     };
-    
+
+    /* GLOBAL FEATURES *************************************************/
+
+    abp.globalFeatures = abp.globalFeatures || {};
+
+    abp.globalFeatures.enabledFeatures = abp.globalFeatures.enabledFeatures || [];
+
+    abp.globalFeatures.isEnabled = function(name){
+        return abp.globalFeatures.enabledFeatures.indexOf(name) != -1;
+    }
+
 })();
