@@ -205,14 +205,16 @@ var abp = abp || {};
         handleAbpErrorResponse: function (jqXHR, userOptions, $dfd) {
             var messagePromise = null;
 
+            var responseJSON = jqXHR.responseJSON ? jqXHR.responseJSON : JSON.parse(jqXHR.responseText);
+
             if (userOptions.abpHandleError !== false) {
-                messagePromise = abp.ajax.showError(jqXHR.responseJSON.error);
+                messagePromise = abp.ajax.showError(responseJSON.error);
             }
 
-            abp.ajax.logError(jqXHR.responseJSON.error);
+            abp.ajax.logError(responseJSON.error);
 
-            $dfd && $dfd.reject(jqXHR.responseJSON.error, jqXHR);
-            userOptions.error && userOptions.error(jqXHR.responseJSON.error, jqXHR);
+            $dfd && $dfd.reject(responseJSON.error, jqXHR);
+            userOptions.error && userOptions.error(responseJSON.error, jqXHR);
 
             if (jqXHR.status === 401 && userOptions.abpHandleError !== false) {
                 abp.ajax.handleUnAuthorizedRequest(messagePromise);
