@@ -15,16 +15,17 @@
 $.extend( $.validator.messages, {
 	required: "تکمیل این فیلد اجباری است.",
 	remote: "لطفا این فیلد را تصحیح کنید.",
-	email: ".لطفا یک ایمیل صحیح وارد کنید",
+	email: "لطفا یک ایمیل صحیح وارد کنید.",
 	url: "لطفا آدرس صحیح وارد کنید.",
-	date: "لطفا یک تاریخ صحیح وارد کنید",
-	dateFA: "لطفا یک تاریخ صحیح وارد کنید",
+	date: "لطفا تاریخ صحیح وارد کنید.",
+	dateFA: "لطفا یک تاریخ صحیح وارد کنید.",
 	dateISO: "لطفا تاریخ صحیح وارد کنید (ISO).",
 	number: "لطفا عدد صحیح وارد کنید.",
-	digits: "لطفا تنها رقم وارد کنید",
+	digits: "لطفا تنها رقم وارد کنید.",
 	creditcard: "لطفا کریدیت کارت صحیح وارد کنید.",
-	equalTo: "لطفا مقدار برابری وارد کنید",
-	extension: "لطفا مقداری وارد کنید که ",
+	equalTo: "لطفا مقدار برابری وارد کنید.",
+	extension: "لطفا مقداری وارد کنید که.",
+	alphanumeric: "لطفا مقدار را عدد (انگلیسی) وارد کنید.",
 	maxlength: $.validator.format( "لطفا بیشتر از {0} حرف وارد نکنید." ),
 	minlength: $.validator.format( "لطفا کمتر از {0} حرف وارد نکنید." ),
 	rangelength: $.validator.format( "لطفا مقداری بین {0} تا {1} حرف وارد کنید." ),
