@@ -14,7 +14,7 @@
  * Locale: GE (Georgian; ქართული)
  */
 $.extend( $.validator.messages, {
-	required: "ეს ველი სავალდებულოა",
+	required: "ეს ველი სავალდებულოა.",
 	remote: "გთხოვთ შეასწოროთ.",
 	email: "გთხოვთ შეიყვანოთ სწორი ფორმატით.",
 	url: "გთხოვთ შეიყვანოთ სწორი ფორმატით.",
