@@ -16,7 +16,7 @@
 $.extend( $.validator.messages, {
 
 	// Core
-	required: "Este campo &eacute; requerido.",
+	required: "Este campo &eacute; obrigat&oacute;rio.",
 	remote: "Por favor, corrija este campo.",
 	email: "Por favor, forne&ccedil;a um endere&ccedil;o de email v&aacute;lido.",
 	url: "Por favor, forne&ccedil;a uma URL v&aacute;lida.",
@@ -32,7 +32,7 @@ $.extend( $.validator.messages, {
 	range: $.validator.format( "Por favor, forne&ccedil;a um valor entre {0} e {1}." ),
 	max: $.validator.format( "Por favor, forne&ccedil;a um valor menor ou igual a {0}." ),
 	min: $.validator.format( "Por favor, forne&ccedil;a um valor maior ou igual a {0}." ),
-	step: $.validator.format( "Por favor, forne&ccedil;a um valor m&acute;tiplo de {0}." ),
+	step: $.validator.format( "Por favor, forne&ccedil;a um valor m&uacute;ltiplo de {0}." ),
 
 	// Metodos Adicionais
 	maxWords: $.validator.format( "Por favor, forne&ccedil;a com {0} palavras ou menos." ),
@@ -62,27 +62,30 @@ $.extend( $.validator.messages, {
 	nieES: "Por favor, forne&ccedil;a um NIE v&aacute;lido.",
 	nifES: "Por favor, forne&ccedil;a um NIF v&aacute;lido.",
 	nowhitespace: "Por favor, n&atilde;o utilize espa&ccedil;os em branco.",
-	pattern: "O formato fornenecido &eacute; inv&aacute;lido.",
-	phoneNL: "Por favor, fornece&ccedil;a um n&uacute;mero de telefone v&aacute;lido.",
-	phoneUK: "Por favor, fornece&ccedil;a um n&uacute;mero de telefone v&aacute;lido.",
-	phoneUS: "Por favor, fornece&ccedil;a um n&uacute;mero de telefone v&aacute;lido.",
-	phonesUK: "Por favor, fornece&ccedil;a um n&uacute;mero de telefone v&aacute;lido.",
-	postalCodeCA: "Por favor, fornece&ccedil;a um n&uacute;mero de c&oacute;digo postal v&aacute;lido.",
-	postalcodeIT: "Por favor, fornece&ccedil;a um n&uacute;mero de c&oacute;digo postal v&aacute;lido.",
-	postalcodeNL: "Por favor, fornece&ccedil;a um n&uacute;mero de c&oacute;digo postal v&aacute;lido.",
-	postcodeUK: "Por favor, fornece&ccedil;a um n&uacute;mero de c&oacute;digo postal v&aacute;lido.",
+	pattern: "O formato fornecido &eacute; inv&aacute;lido.",
+	phoneNL: "Por favor, forne&ccedil;a um n&uacute;mero de telefone v&aacute;lido.",
+	phoneUK: "Por favor, forne&ccedil;a um n&uacute;mero de telefone v&aacute;lido.",
+	phoneUS: "Por favor, forne&ccedil;a um n&uacute;mero de telefone v&aacute;lido.",
+	phonesUK: "Por favor, forne&ccedil;a um n&uacute;mero de telefone v&aacute;lido.",
+	postalCodeCA: "Por favor, forne&ccedil;a um n&uacute;mero de c&oacute;digo postal v&aacute;lido.",
+	postalcodeIT: "Por favor, forne&ccedil;a um n&uacute;mero de c&oacute;digo postal v&aacute;lido.",
+	postalcodeNL: "Por favor, forne&ccedil;a um n&uacute;mero de c&oacute;digo postal v&aacute;lido.",
+	postcodeUK: "Por favor, forne&ccedil;a um n&uacute;mero de c&oacute;digo postal v&aacute;lido.",
 	postalcodeBR: "Por favor, forne&ccedil;a um CEP v&aacute;lido.",
 	require_from_group: $.validator.format( "Por favor, forne&ccedil;a pelo menos {0} destes campos." ),
 	skip_or_fill_minimum: $.validator.format( "Por favor, optar entre ignorar esses campos ou preencher pelo menos {0} deles." ),
 	stateUS: "Por favor, forne&ccedil;a um estado v&aacute;lido.",
 	strippedminlength: $.validator.format( "Por favor, forne&ccedil;a pelo menos {0} caracteres." ),
-	time: "Por favor, forne&ccedil;a um hor&aacute;rio v&aacute;lido, no intervado de 00:00 e 23:59.",
-	time12h: "Por favor, forne&ccedil;a um hor&aacute;rio v&aacute;lido, no intervado de 01:00 e 12:59 am/pm.",
-	url2: "Por favor, fornece&ccedil;a uma URL v&aacute;lida.",
-	vinUS: "O n&uacute;mero de identifica&ccedil;&atilde;o de ve&iacute;culo informada (VIN) &eacute; inv&aacute;lido.",
-	zipcodeUS: "Por favor, fornece&ccedil;a um c&oacute;digo postal americano v&aacute;lido.",
-	ziprange: "O c&oacute;digo postal deve estar entre 902xx-xxxx e 905xx-xxxx",
-	cpfBR: "Por favor, forne&ccedil;a um CPF v&aacute;lido."
+	time: "Por favor, forne&ccedil;a um hor&aacute;rio v&aacute;lido, no intervado de 00:00 a 23:59.",
+	time12h: "Por favor, forne&ccedil;a um hor&aacute;rio v&aacute;lido, no intervado de 01:00 a 12:59 am/pm.",
+	url2: "Por favor, forne&ccedil;a uma URL v&aacute;lida.",
+	vinUS: "O n&uacute;mero de identifica&ccedil;&atilde;o de ve&iacute;culo informado (VIN) &eacute; inv&aacute;lido.",
+	zipcodeUS: "Por favor, forne&ccedil;a um c&oacute;digo postal americano v&aacute;lido.",
+	ziprange: "O c&oacute;digo postal deve estar entre 902xx-xxxx e 905xx-xxxx.",
+	cpfBR: "Por favor, forne&ccedil;a um CPF v&aacute;lido.",
+	nisBR: "Por favor, forne&ccedil;a um NIS/PIS v&aacute;lido.",
+	cnhBR: "Por favor, forne&ccedil;a um CNH v&aacute;lido.",
+	cnpjBR: "Por favor, forne&ccedil;a um CNPJ v&aacute;lido."
 } );
 return $;
 }));
