@@ -29,7 +29,8 @@ $.extend( $.validator.messages, {
 	rangelength: $.validator.format( "Унесите вредност дугачку између {0} и {1} карактера." ),
 	range: $.validator.format( "Унесите вредност између {0} и {1}." ),
 	max: $.validator.format( "Унесите вредност мању или једнаку {0}." ),
-	min: $.validator.format( "Унесите вредност већу или једнаку {0}." )
+	min: $.validator.format( "Унесите вредност већу или једнаку {0}." ),
+	step: $.validator.format( "Унесите вредност која је умножак броја {0}." )
 } );
 return $;
 }));
