@@ -14,6 +14,7 @@
  */
 $.extend( $.validator.messages, {
 	required: "Detta f&auml;lt &auml;r obligatoriskt.",
+	remote: "Var snäll och åtgärda detta fält.",
 	maxlength: $.validator.format( "Du f&aring;r ange h&ouml;gst {0} tecken." ),
 	minlength: $.validator.format( "Du m&aring;ste ange minst {0} tecken." ),
 	rangelength: $.validator.format( "Ange minst {0} och max {1} tecken." ),
@@ -27,7 +28,8 @@ $.extend( $.validator.messages, {
 	range: $.validator.format( "Ange ett v&auml;rde mellan {0} och {1}." ),
 	max: $.validator.format( "Ange ett v&auml;rde som &auml;r mindre eller lika med {0}." ),
 	min: $.validator.format( "Ange ett v&auml;rde som &auml;r st&ouml;rre eller lika med {0}." ),
-	creditcard: "Ange ett korrekt kreditkortsnummer."
+	creditcard: "Ange ett korrekt kreditkortsnummer.",
+	pattern: "Ogiltigt format."
 } );
 return $;
 }));
