@@ -23,7 +23,7 @@ $.extend( $.validator.messages, {
 	digits: "Илтимос, танҳо рақам ворид кунед.",
 	creditcard: "Илтимос, кредит карди саҳеҳ ворид кунед.",
 	equalTo: "Илтимос, миқдори баробар ворид кунед.",
-	extension: "Илтимос, қофияи файлро дуруст интихоб кунед",
+	extension: "Илтимос, қофияи файлро дуруст интихоб кунед.",
 	maxlength: $.validator.format( "Илтимос, бештар аз {0} рамз ворид накунед." ),
 	minlength: $.validator.format( "Илтимос, камтар аз {0} рамз ворид накунед." ),
 	rangelength: $.validator.format( "Илтимос, камтар аз {0} ва зиёда аз {1} рамз ворид кунед." ),
