@@ -29,7 +29,7 @@ $.extend( $.validator.messages, {
 	rangelength: $.validator.format( "Hãy nhập từ {0} đến {1} kí tự." ),
 	range: $.validator.format( "Hãy nhập từ {0} đến {1}." ),
 	max: $.validator.format( "Hãy nhập từ {0} trở xuống." ),
-	min: $.validator.format( "Hãy nhập từ {1} trở lên." )
+	min: $.validator.format( "Hãy nhập từ {0} trở lên." )
 } );
 return $;
 }));
